+++ lib/everest/everest_api_types/include/everest_api_types/energy/API.hpp
@@ -12,7 +12,7 @@
 namespace everest::lib::API::V1_0::types::energy {
 
 struct NumberWithSource {
-    float value;
+    double value;
     std::string source;
 };
 
@@ -22,21 +22,21 @@ struct IntegerWithSource {
 };
 
 struct FrequencyWattPoint {
-    float frequency_Hz;
-    float total_power_W;
+    double frequency_Hz;
+    double total_power_W;
 };
 
 struct SetpointType {
     int32_t priority;
     std::string source;
-    std::optional<float> ac_current_A;
-    std::optional<float> total_power_W;
+    std::optional<double> ac_current_A;
+    std::optional<double> total_power_W;
     std::optional<std::vector<types::energy::FrequencyWattPoint>> frequency_table;
 };
 
 struct PricePerkWh {
     std::string timestamp;
-    float value;
+    double value;
     std::string currency;
 };
 
@@ -60,7 +60,7 @@ struct ScheduleReqEntry {
     std::string timestamp;
     LimitsReq limits_to_root;
     LimitsReq limits_to_leaves;
-    std::optional<float> conversion_efficiency;
+    std::optional<double> conversion_efficiency;
     std::optional<PricePerkWh> price_per_kwh;
 };
 
@@ -89,10 +89,10 @@ struct EnforcedLimits {
 };
 
 struct CapabilityLimits {
-    float max_current_A;
+    double max_current_A;
     int32_t max_phase_count;
-    float nominal_voltage_V;
-    float total_power_W;
+    double nominal_voltage_V;
+    double total_power_W;
 };
 
 } // namespace everest::lib::API::V1_0::types::energy
