+++ lib/everest/everest_api_types/include/everest_api_types/power_supply_DC/API.hpp
@@ -10,30 +10,30 @@ namespace everest::lib::API::V1_0::types::power_supply_DC {
 
 struct Capabilities {
     bool bidirectional;
-    float current_regulation_tolerance_A;
-    float peak_current_ripple_A;
-    float max_export_voltage_V;
-    std::optional<float> nominal_max_export_voltage_V;
-    float min_export_voltage_V;
-    std::optional<float> nominal_min_export_voltage_V;
-    float max_export_current_A;
-    std::optional<float> nominal_max_export_current_A;
-    float min_export_current_A;
-    std::optional<float> nominal_min_export_current_A;
-    float max_export_power_W;
-    std::optional<float> nominal_max_export_power_W;
-    std::optional<float> max_import_voltage_V;
-    std::optional<float> nominal_max_import_voltage_V;
-    std::optional<float> min_import_voltage_V;
-    std::optional<float> nominal_min_import_voltage_V;
-    std::optional<float> max_import_current_A;
-    std::optional<float> nominal_max_import_current_A;
-    std::optional<float> min_import_current_A;
-    std::optional<float> nominal_min_import_current_A;
-    std::optional<float> max_import_power_W;
-    std::optional<float> nominal_max_import_power_W;
-    std::optional<float> conversion_efficiency_import;
-    std::optional<float> conversion_efficiency_export;
+    double current_regulation_tolerance_A;
+    double peak_current_ripple_A;
+    double max_export_voltage_V;
+    std::optional<double> nominal_max_export_voltage_V;
+    double min_export_voltage_V;
+    std::optional<double> nominal_min_export_voltage_V;
+    double max_export_current_A;
+    std::optional<double> nominal_max_export_current_A;
+    double min_export_current_A;
+    std::optional<double> nominal_min_export_current_A;
+    double max_export_power_W;
+    std::optional<double> nominal_max_export_power_W;
+    std::optional<double> max_import_voltage_V;
+    std::optional<double> nominal_max_import_voltage_V;
+    std::optional<double> min_import_voltage_V;
+    std::optional<double> nominal_min_import_voltage_V;
+    std::optional<double> max_import_current_A;
+    std::optional<double> nominal_max_import_current_A;
+    std::optional<double> min_import_current_A;
+    std::optional<double> nominal_min_import_current_A;
+    std::optional<double> max_import_power_W;
+    std::optional<double> nominal_max_import_power_W;
+    std::optional<double> conversion_efficiency_import;
+    std::optional<double> conversion_efficiency_export;
 };
 
 enum class Mode {
@@ -56,8 +56,8 @@ struct ModeRequest {
 };
 
 struct VoltageCurrent {
-    float voltage_V; ///< Voltage in V
-    float current_A; ///< Current in A
+    double voltage_V; ///< Voltage in V
+    double current_A; ///< Current in A
 };
 
 enum class ErrorEnum {
